Add RPC encoding module (#13)
`Exiris.Rpc.Encoding` is responsible for enconding requests into json
and decoding responses from json. This is needed so we don't have
coupled things in the app.

We also included a `:encoder` and a `:decoder` in the `Provider`
structure so we can pass it to the transport layer that can use it to
encode and decode structures.

This is parametrized and decoupled.