Add support for dynamic configuration in the Provider (#56)
**Why:**

We need to be able to configure the Providers via app config that might
be simple environment configuration or just runtime configuration. We
need to account for those cases.
That's needed because we might want to hide rpc_urls and have more
private configs that do not need to be compiled.

**How:**

By:
- removing `Provider.Validator` macros, given they were making
compile-time checks and are not useful anymore.
- moving `Provider.Generator` macros to the `Provider` module, given
that we do not have separation of concerns anymore because we don't have
validations;
- updating `Provider.get_client` to fetch `Application` config and merge
it with compile-time config. Compile-time configuration takes precedence
over the dynamic config;
- adding tests that test the new Provider configuration behaviours;