Add RPC Call struct for pipeable operations in the Client (#59)
**Why:**

We need a more fluent way to make requests with the RPC client.

**How:**

By returning an `Rpc.Call` struct when creating a new request and
passing it to other requests.
We can then use this `Rpc.Call` as the argument to the `Client.send/1`
function.

An example of a single request would be:
```elixir
client 
|> Rpc.request("eth_blockNumber", []) 
|> Rpc.send()
```

A batch request would be something like:
```elixir
client 
|> Rpc.request("eth_blockNumber", []) 
|> Rpc.request("eth_getBalance", ["0x742d...", "latest"])
|> Rpc.send()
```

If we still want to make a raw request for a specific reason, we can do
something like:
```elixir
request = Rpc.raw_request("eth_blockNumber", [])
Rpc.send(request, client)
# or
Rpc.send(client, request)
```

This offers a simpler and more fluent api for making requests.