Fix - :otp_app is not require in the Provider (#60)
**Why:**

In `Exth.Provider` we fetched the module `:otp_app` with
`Keyword.fetch!/2`.
The thing is that if the module didn't have this option, it would raise.
But this is not desired because we use `:otp_app` to later fetch the app
config, which might not be required if all config is set as options in
the provider.

**How:**

By replacing `Keyword.fetch!/2` with `Keyword.get/3`, which defaults to
`nil` when there's no `:otp_app`.