Transport layer is only responsible for sending/receiving requests (#61)
**Why:**

We want the core parts of `Exth` to have clear responsibilities and to
be as minimal as possible.
So far, the `Transport` layer is responsible for sending/receiving
requests as well as serializing/deserializing `Request`s and
`Response`s.
The last part should not be within the Transport scope.

**How:**

By moving encoding/decoding to `Client` and making `Transport` layer
simpler.
As a consequence, we needed to improve `Client.send/2` logic.