Add InnerClient as foundation for Async Transports (#63)
**Why:**

Soon we'll add Websocket support, which is an async transport.
Async transports have a peculiarity that makes them very different from
sync transports like HTTP. Websockets are bi-directional, and we can use
the same connection to serve concurrent requests, meaning that the
request management is complex, and we need to keep a map of the requests
to match the received responses later.

**How:**

By adding the concept of an `InnerClient`, which is a Genserver
responsible for matching requests and responses.
The `InnerClient` holds an `ETS` table where its keys are the request
IDs, and the values are the PIDs of the processes waiting on a response
for those specific requests.