Replace InnerClient with MessageHandler (#65)
**Why:**

So far, we have had the concept of an `InnerClient` within a `Client`.
The `InnerClient` was a GenServer that used an ETS table to store the
mappings between requests and its caller's PID.
This approach seemed to be overly complex for what it was, and it
required more attention because there were a lot of things done
manually.

**How:**

Elixir already has the concept of a `Registry`, which is a decentralized
and scalable key-value process storage.
In this PR, we replace the `InnerClient` with a Registry, which we call
`MessageHandler`. There's a mapping 1 to 1 between a client and a
`MessageHandler`.
This Registry registers the caller process within the request_id key.
Later, when there's a response, there's a lookup on the Registry for the
matching request_id, and the response is delivered to the caller.