Add support for subscriptions over Websockets (#67)
**Why:**

Unlike HTTP, with WebSockets, we don't need to continuously make
requests when we want specific information. WebSockets maintain a
network connection, so we can listen for changes.
When connected through a WebSocket, we may use two additional methods:
`eth_subscribe` and `eth_unsubscribe`. These methods will allow us to
listen for particular events and be notified immediately.

**How:**

By making the distinction between normal RPCs and `eth_subscribe` and
`eth_unsubscribe` in the `MessageHandler`.
After receiving a subscription response, we register the subscription ID
under the caller process, and then we send `SubscriptionEvents` to this
process.

Solves: #66