Add Supervision Tree for Transport layer (#76)
**Why:**

So far, we have had no supervision tree for Exth and no registries for
its processes.
We need process registries so that we can address processes by name,
rather than by PID. This is important because if a process crashes and
its supervisor spawns another one, we should be able to find the new
process by the same name.
On the other side, we need supervisors so we can create processes under
them, and we can have strategies for recovering from crashes. This will
make Exth more resilient for sure.
We need to tackle these problems in phases, and we should start with the
Transport layer.

**How:**

By:
- adding a `Transport.Registry` that will be responsible for registering
all the named processes within the Transport layer;
- adding a `Transport.Websocket.DynamicSupervisor`, that will be
responsible for creating `Transport.Websocket` processes on demand.
- adding a `Transport.Supervisor` that will be the top supervisor for
the Transport layer and will initialize `Transport.Registry` and
`Transport.Websocket.DynamicSupervisor`;

```mermaid
flowchart TD
      A["Exth.Supervisor :one_for_one"] --> B["Transport.Supervisor :one_for_one"]
      B --> C["Transport.Registry"]
      B --> D["Transport.Websocket.DynamicSupervisor :one_for_one"]
      C -.registers.- E["Websocket"]
      D --> E
      E@{ shape: procs}
```