Transport - use Websocket process via-tuple name instead of its PID (#77)
**Why:**

Processes can die and be restarted. Every time they're restarted, they
get a new PID.
Our `Transport.Websocket` struct would keep track of the process PID
which is not the most resilient solution.

**How:**

Given the process is also registered under a specific name under a
Registry, we now use that name in the struct instead of the PID.